Fix the return type of c_missing_value_flt (#651)
* Added a (failing) test of the Fortran interface to missingValue().

* Fixed type of value returned by c_missing_value_flt().

+++ src/oops/util/missing_values_mod.F90
@@ -19,7 +19,7 @@ public missing_value
 !-------------------------------------------------------------------------------
 interface
 !-------------------------------------------------------------------------------
-real(kind=c_double) function c_missing_value_flt() bind(C,name='missing_value_flt_f')
+real(kind=c_float) function c_missing_value_flt() bind(C,name='missing_value_flt_f')
   use, intrinsic :: iso_c_binding
   implicit none
 end function c_missing_value_flt
